Refactoring the structure to isolate the model, the UI calls and the IO hardcoded blochain management. Tested with node_command_view.
