+++ .gitignore
@@ -2,3 +2,4 @@
 /resources/
 /__*/
 *__pycache__
+/venv/
