+++ blockchain_model/transaction.py
@@ -3,12 +3,12 @@ import time as tm
 
 class Transaction:
 
-    def __init__(self, tx_recipient, tx_sender=None, tx_amount=0, tx_time=None, tx_signature=None):
+    def __init__(self, tx_sender, tx_recipient, tx_amount, tx_signature, tx_time=None):
         self.__recipient = tx_recipient
-        self.__sender = tx_sender if tx_sender is not None else 'System0'
+        self.__sender = tx_sender
         self.__amount = tx_amount
+        self.__signature = tx_signature
         self.__timestamp = tx_time if tx_time is not None else tm.time()
-        self.__signature = tx_signature if tx_signature is not None else 'MINING'
 
     @property
     def sender(self):
