+++ requirements.txt
@@ -1,0 +1,17 @@
+certifi==2019.11.28
+cffi==1.13.2
+Click==7.0
+Flask==1.1.1
+Flask-Cors==3.0.8
+Flask-Login==0.4.1
+Flask-SQLAlchemy==2.4.1
+itsdangerous==1.1.0
+Jinja2==2.10.3
+MarkupSafe==1.1.1
+pycparser==2.19
+pycryptodomex==3.9.4
+six==1.13.0
+SQLAlchemy==1.3.12
+Werkzeug==0.16.0
+wincertstore==0.2
+requests==2.22.0
